- changed path of componenent directory: src/assets -> src/ - renamed function onEdit/onDelete to onEditClick/onEditClick in TodoItem/TodoListBox.
